requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
python-dotenv
anthropic
google-genai
openai
